pyproject.toml
[project]
name = "docchat-git"
version = "0.1.0"
description = "Docchat agentic AI"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "accelerate==1.12.0",
    "aiofiles==23.2.1",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.2",
    "aiosignal==1.4.0",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.0",
    "appnope==0.1.4",
    "asgiref==3.11.0",
    "asttokens==3.0.1",
    "attrs==25.4.0",
    "autoflake==2.3.1",
    "backoff==2.2.1",
    "bcrypt==5.0.0",
    "beautifulsoup4==4.14.3",
    "black==25.12.0",
    "brotli==1.2.0",
    "build==1.3.0",
    "cachetools==6.2.4",
    "certifi==2025.11.12",
    "charset-normalizer==3.4.4",
    "chroma-hnswlib==0.7.6",
    "chromadb==0.6.3",
    "click==8.3.1",
    "coloredlogs==15.0.1",
    "comm==0.2.3",
    "contourpy==1.3.3",
    "cycler==0.12.1",
    "dataclasses-json==0.6.7",
    "debugpy==1.8.19",
    "decorator==5.2.1",
    "deepsearch-glm==1.0.0",
    "dill==0.4.0",
    "distro==1.9.0",
    "docling==2.15.0",
    "docling-core==2.57.0",
    "docling-parse==3.4.0",
    "durationpy==0.10",
    "easyocr==1.7.2",
    "et-xmlfile==2.0.0",
    "executing==2.2.1",
    "fastapi==0.125.0",
    "fastcore==1.10.0",
    "ffmpy==1.0.0",
    "filelock==3.20.1",
    "filetype==1.2.0",
    "flatbuffers==25.12.19",
    "fonttools==4.61.1",
    "frozenlist==1.8.0",
    "fsspec==2025.12.0",
    "google-auth==2.45.0",
    "googleapis-common-protos==1.72.0",
    "gradio==6.2.0",
    "gradio-client==2.0.2",
    "greenlet==3.3.0",
    "groovy==0.1.2",
    "grpcio==1.76.0",
    "h11==0.16.0",
    "hf-xet==1.2.0",
    "httpcore==1.0.9",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "httpx-sse==0.4.3",
    "huggingface-hub==0.36.0",
    "humanfriendly==10.0",
    "idna==3.11",
    "imageio==2.37.2",
    "importlib-metadata==8.7.0",
    "importlib-resources==6.5.2",
    "iniconfig==2.3.0",
    "ipykernel==7.1.0",
    "ipython==9.8.0",
    "ipython-pygments-lexers==1.1.1",
    "isort==7.0.0",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "jiter==0.12.0",
    "jmespath==1.0.1",
    "jsonlines==4.0.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonref==1.1.0",
    "jsonschema==4.25.1",
    "jsonschema-specifications==2025.9.1",
    "jupyter-client==8.7.0",
    "jupyter-core==5.9.1",
    "kiwisolver==1.4.9",
    "kubernetes==34.1.0",
    "langchain==0.3.27",
    "langchain-community==0.3.16",
    "langchain-core==0.3.80",
    "langchain-openai==0.3.2",
    "langchain-text-splitters==0.3.11",
    "langgraph==0.2.68",
    "langgraph-checkpoint==2.1.2",
    "langgraph-sdk==0.1.74",
    "langsmith==0.3.45",
    "latex2mathml==3.78.1",
    "lazy-loader==0.4",
    "loguru==0.7.3",
    "lomond==0.3.3",
    "lxml==5.4.0",
    "markdown-it-py==4.0.0",
    "marko==2.2.1",
    "markupsafe==2.1.5",
    "marshmallow==3.26.1",
    "matplotlib==3.10.8",
    "matplotlib-inline==0.2.1",
    "mdurl==0.1.2",
    "mmh3==5.2.0",
    "mpire==2.10.2",
    "mpmath==1.3.0",
    "multidict==6.7.0",
    "multiprocess==0.70.18",
    "mypy-extensions==1.1.0",
    "nest-asyncio==1.6.0",
    "networkx==3.6.1",
    "ninja==1.13.0",
    "numpy==1.26.4",
    "oauthlib==3.3.1",
    "obiter==0.0.4",
    "openai==1.109.1",
    "opencv-python-headless==4.10.0.84",
    "openpyxl==3.1.5",
    "opentelemetry-api==1.39.1",
    "opentelemetry-exporter-otlp-proto-common==1.39.1",
    "opentelemetry-exporter-otlp-proto-grpc==1.39.1",
    "opentelemetry-instrumentation==0.60b1",
    "opentelemetry-instrumentation-asgi==0.60b1",
    "opentelemetry-instrumentation-fastapi==0.60b1",
    "opentelemetry-proto==1.39.1",
    "opentelemetry-sdk==1.39.1",
    "opentelemetry-semantic-conventions==0.60b1",
    "opentelemetry-util-http==0.60b1",
    "orjson==3.11.5",
    "ormsgpack==1.12.1",
    "overrides==7.7.0",
    "packaging==25.0",
    "pandas==2.3.3",
    "parso==0.8.5",
    "pathspec==0.12.1",
    "pexpect==4.9.0",
    "pillow==11.3.0",
    "platformdirs==4.5.1",
    "playwright==1.57.0",
    "pluggy==1.6.0",
    "posthog==5.4.0",
    "prompt-toolkit==3.0.52",
    "propcache==0.4.1",
    "protobuf==6.33.2",
    "psutil==7.2.1",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pybase64==1.4.3",
    "pyclipper==1.4.0",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-settings==2.12.0",
    "pydub==0.25.1",
    "pyee==13.0.0",
    "pyflakes==3.4.0",
    "pygments==2.19.2",
    "pyparsing==3.3.1",
    "pypdf==5.2.0",
    "pypdfium2==4.30.0",
    "pypika==0.48.9",
    "pyproject-hooks==1.2.0",
    "pytest==9.0.2",
    "python-bidi==0.6.7",
    "python-dateutil==2.9.0.post0",
    "python-docx==1.2.0",
    "python-dotenv==1.2.1",
    "python-multipart==0.0.21",
    "python-pptx==1.0.2",
    "pytokens==0.3.0",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "pyzmq==27.1.0",
    "rank-bm25==0.2.2",
    "referencing==0.37.0",
    "regex==2025.11.3",
    "requests==2.32.3",
    "requests-oauthlib==2.0.0",
    "requests-toolbelt==1.0.0",
    "rich==14.2.0",
    "rpds-py==0.30.0",
    "rsa==4.9.1",
    "rtree==1.4.1",
    "ruff==0.14.10",
    "safehttpx==0.1.7",
    "safetensors==0.7.0",
    "scikit-image==0.25.2",
    "scipy==1.16.3",
    "seaborn==0.13.2",
    "semantic-version==2.10.0",
    "semchunk==2.2.2",
    "shapely==2.1.2",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.8.1",
    "sqlalchemy==2.0.45",
    "stack-data==0.6.3",
    "starlette==0.50.0",
    "sympy==1.14.0",
    "tabulate==0.9.0",
    "tenacity==9.1.2",
    "tifffile==2025.12.12",
    "tiktoken==0.12.0",
    "tokenizers==0.22.1",
    "tomlkit==0.13.3",
    "torch==2.2.2",
    "torchvision==0.17.2",
    "tornado==6.5.4",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "transformers==4.57.3",
    "tree-sitter==0.25.2",
    "tree-sitter-c==0.24.1",
    "tree-sitter-java==0.23.5",
    "tree-sitter-javascript==0.25.0",
    "tree-sitter-python==0.25.0",
    "tree-sitter-typescript==0.23.2",
    "typer==0.12.5",
    "typer-slim==0.20.1",
    "typing-extensions==4.15.0",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.3",
    "urllib3==2.3.0",
    "uuid-utils==0.12.0",
    "uvicorn==0.38.0",
    "uvloop==0.22.1",
    "watchfiles==1.1.1",
    "wcwidth==0.2.14",
    "websocket-client==1.9.0",
    "websockets==14.2",
    "wrapt==1.17.3",
    "xlsxwriter==3.2.9",
    "yarl==1.22.0",
    "zipp==3.23.0",
    "zstandard==0.23.0",
]
